Rename file to known name to allow for namespaced extension name (#77)
* Rename file to known name to allow for namespaced extension name

* fix tstargets
